Parsing refactor almost finish
Functions parse_plane and parse_cylinder are refactor.
Add of parse_direction for parsing a vector, check is range and normalize it.
Add of parse_float witch is a rename of parse_radius.
Change of parse_camera and parse_sphere for use parse_direction adn parse_float

Modify functions in print.c and draw.c for compile the project.

We have some error message like 'X file is modify of 0.001s in the future' in compilation.
I have no idea of where is the error, but the project compile and work.

I have also modify the SRCS to have an explicite path of files. You can comment it and uncomment
the previous SRCS line.