Ambiant light and light menu ok
I've done many 'utils' functions for display things on the menu.
Like a float or the color.

I've done an enum 'draw_state' for know when we need to put image (NEED_REDRAW),
when we need to draw text (DRAW_TEXT) and when we draw only the pixels (DRAW_PIXELS)

The menu of the ambiant light and the light work like intended.

Sometimes lag can appear. I'll check this tomorrow.

The intensity of the light doesn't work when it's over 0.0 (It like a switch and not smouth)